latest

+++ build/module-builder/src/build.ts
@@ -1,4 +1,5 @@
 import * as babel from '@babel/core'
+import chalk from 'chalk'
 import fs from 'fs'
 import fse from 'fs-extra'
 import glob from 'glob'
@@ -6,6 +7,7 @@ import mkdirp from 'mkdirp'
 import os from 'os'
 import path from 'path'
 import rimraf from 'rimraf'
+import * as ts from 'typescript'
 import { generateSchema, getProgramFromFiles } from 'typescript-json-schema'
 
 import { debug, error, normal } from './log'
@@ -22,6 +24,8 @@ export default async (argv: any) => {
 }
 
 export async function buildBackend(modulePath: string) {
+  const start = Date.now()
+
   let babelConfig: babel.TransformOptions = {
     presets: [
       [
@@ -35,14 +39,15 @@ export async function buildBackend(modulePath: string) {
       '@babel/preset-typescript',
       '@babel/preset-react'
     ],
-    sourceMaps: 'both',
+
+    sourceMaps: true,
     sourceRoot: path.join(modulePath, 'src/backend'),
     parserOpts: {
       allowReturnOutsideFunction: true
     },
     plugins: ['@babel/plugin-proposal-class-properties', '@babel/plugin-proposal-function-bind'],
     sourceType: 'module',
-    cwd: modulePath
+    cwd: path.resolve(__dirname, '..')
   }
 
   const babelFile = path.join(modulePath, 'babel.backend.js')
@@ -52,47 +57,79 @@ export async function buildBackend(modulePath: string) {
     babelConfig = require(babelFile)(babelConfig)
   }
 
-  const files = glob.sync('src/**/*.+(ts|js|jsx|tsx)', {
+  const tsConfigFile = ts.findConfigFile(modulePath, ts.sys.fileExists, 'tsconfig.json')
+  const skipCheck = process.argv.find(x => x.toLowerCase() === '--skip-check')
+
+  let validCode = true
+  if (!skipCheck && tsConfigFile) {
+    validCode = runTypeChecker(modulePath)
+  }
+
+  if (validCode) {
+    rimraf.sync(path.join(modulePath, 'dist'))
+
+    copyExtraFiles(modulePath)
+    compileBackend(modulePath, babelConfig)
+
+    normal(`Generated backend (${Date.now() - start} ms)`, path.basename(modulePath))
+  }
+}
+
+// Allows to copy additional files to the dist directory of the module
+const copyExtraFiles = (modulePath: string) => {
+  const extrasFile = path.join(modulePath, 'build.extras.js')
+  if (!fs.existsSync(extrasFile)) {
+    return
+  }
+
+  const extras = require(extrasFile)
+  if (extras && extras.copyFiles) {
+    for (const instruction of extras.copyFiles) {
+      const toCopy = glob.sync(instruction, {
+        cwd: modulePath,
+        dot: true
+      })
+
+      for (const file of toCopy) {
+        const fromFull = path.join(modulePath, file)
+        const dest = file.replace(/^src\//i, 'dist/').replace(/\.ts$/i, '.js')
+        const destFull = path.join(modulePath, dest)
+        mkdirp.sync(path.dirname(destFull))
+        fse.copySync(fromFull, destFull)
+        debug(`Copied "${file}" -> "${dest}"`)
+      }
+    }
+  }
+}
+
+const compileBackend = (modulePath: string, babelConfig) => {
+  const files = glob.sync('src/**/*.+(ts|js|jsx|tsx|json)', {
     cwd: modulePath,
     dot: true,
     ignore: ['**/*.d.ts', '**/views/**/*.*', '**/config.ts']
   })
 
-  rimraf.sync(path.join(modulePath, 'dist'))
-
-  // Allows to copy additional files to the dist directory of the module
-  const extrasFile = path.join(modulePath, 'build.extras.js')
-  if (fs.existsSync(extrasFile)) {
-    const extras = require(extrasFile)
-    if (extras && extras.copyFiles) {
-      for (const instruction of extras.copyFiles) {
-        const toCopy = glob.sync(instruction, {
-          cwd: modulePath,
-          dot: true
-        })
-
-        for (const file of toCopy) {
-          const fromFull = path.join(modulePath, file)
-          const dest = file.replace(/^src\//i, 'dist/').replace(/\.ts$/i, '.js')
-          const destFull = path.join(modulePath, dest)
-          mkdirp.sync(path.dirname(destFull))
-          fse.copySync(fromFull, destFull)
-          debug(`Copied "${file}" -> "${dest}"`)
-        }
-      }
-    }
-  }
-
+  const copyWithoutTransform = ['actions', 'hooks', 'examples', 'content-types']
   const outputFiles = []
 
   for (const file of files) {
+    const dest = file.replace(/^src\//i, 'dist/').replace(/\.ts$/i, '.js')
+    mkdirp.sync(path.dirname(dest))
+
+    if (copyWithoutTransform.find(x => file.startsWith(`src/${x}`)) || file.endsWith('.json')) {
+      fs.writeFileSync(dest, fs.readFileSync(`${modulePath}/${file}`, 'utf8'))
+      continue
+    }
+
     try {
       const dBefore = Date.now()
       const result = babel.transformFileSync(file, babelConfig)
+      const destMap = dest + '.map'
 
-      const dest = file.replace(/^src\//i, 'dist/').replace(/.ts$/i, '.js')
-      mkdirp.sync(path.dirname(dest))
-      fs.writeFileSync(dest, result.code)
+      fs.writeFileSync(dest, result.code + os.EOL + `//# sourceMappingURL=${path.basename(destMap)}`)
+      result.map.sources = [path.relative(babelConfig.sourceRoot, file)]
+      fs.writeFileSync(destMap, JSON.stringify(result.map))
+
       const totalTime = Date.now() - dBefore
 
       debug(`Generated "${dest}" (${totalTime} ms)`)
@@ -105,7 +142,7 @@ export async function buildBackend(modulePath: string) {
   }
 }
 
-export async function buildConfigSchema(modulePath: string) {
+const buildConfigSchema = async (modulePath: string) => {
   const config = path.resolve(modulePath, 'src', 'config.ts')
   if (!fs.existsSync(config)) {
     return
@@ -129,4 +166,55 @@ export async function buildConfigSchema(modulePath: string) {
 
   mkdirp.sync(path.resolve(modulePath, 'assets'))
   fs.writeFileSync(path.resolve(modulePath, 'assets', 'config.schema.json'), schema)
+}
+
+const getTsConfig = (rootFolder: string): ts.ParsedCommandLine => {
+  const parseConfigHost: ts.ParseConfigHost = {
+    fileExists: ts.sys.fileExists,
+    readFile: ts.sys.readFile,
+    readDirectory: ts.sys.readDirectory,
+    useCaseSensitiveFileNames: true
+  }
+
+  const configFileName = ts.findConfigFile(rootFolder, ts.sys.fileExists, 'tsconfig.json')
+  const { config } = ts.readConfigFile(configFileName, ts.sys.readFile)
+
+  // These 3 objects are identical for all modules, but can't be in tsconfig.shared because the root folder is not processed correctly
+  const fixedModuleConfig = {
+    ...config,
+    compilerOptions: {
+      ...config.compilerOptions,
+      typeRoots: ['./node_modules/@types', './node_modules', './src/typings']
+    },
+    exclude: ['**/*.test.ts', './src/views/**', '**/node_modules/**'],
+    include: ['../../src/typings/*.d.ts', '**/*.ts']
+  }
+
+  return ts.parseJsonConfigFileContent(fixedModuleConfig, parseConfigHost, rootFolder)
+}
+
+const runTypeChecker = (rootFolder: string): boolean => {
+  const { options, fileNames } = getTsConfig(rootFolder)
+
+  const program = ts.createProgram(fileNames, options)
+  const diagnostics = ts.getPreEmitDiagnostics(program).concat(program.emit().diagnostics)
+
+  for (const { file, start, messageText, code } of diagnostics) {
+    if (file) {
+      const { line, character } = file.getLineAndCharacterOfPosition(start!)
+      const message = ts.flattenDiagnosticMessageText(messageText, '\n')
+
+      error(
+        chalk.bold(
+          `[ERROR] in ${chalk.cyan(file.fileName)} (${line + 1},${character + 1})
+                 TS${code}: ${message}
+                 `
+        )
+      )
+    } else {
+      error(ts.flattenDiagnosticMessageText(messageText, '\n'))
+    }
+  }
+
+  return !diagnostics.length
 }
